RadProfile line cleanup
