+++ src/eradiate/radprops/_us76_approx.py
@@ -12,6 +12,7 @@ from ._util_mono import get_us76_u86_4_spectrum_filename
 from .absorption import compute_sigma_a
 from .rayleigh import compute_sigma_s_air
 from .. import data
+from .._mode import UnsupportedModeError
 from ..attrs import documented, parse_docs
 from ..ckd import Bindex
 from ..thermoprops import us76
@@ -37,12 +38,13 @@ class US76ApproxRadProfile(RadProfile):
     atmospheric profile based on the original U.S. Standard Atmosphere 1976
     atmosphere model.
 
-    .. note::
+    Warnings
+    --------
+    This class does not support ``ckd`` modes.
 
-       Instantiating this class requires to download the absorption dataset
-       ``spectra-us76_u86_4`` and place it in ``$ERADIATE_DIR/resources/data/``.
-
-    The :mod:`~eradiate.thermoprops.us76` module implements the original *U.S.
+    Notes
+    -----
+    * The :mod:`~eradiate.thermoprops.us76` module implements the original *U.S.
     Standard Atmosphere 1976* atmosphere model, as defined by the
     :cite:`NASA1976USStandardAtmosphere` technical report.
     In the original atmosphere model, the gases are assumed well-mixed below
@@ -62,50 +64,48 @@ class US76ApproxRadProfile(RadProfile):
     This approximation assumes that the absorption coefficient does not vary
     much whether the mixing ratios of the absorbing gas mixture are those
     below or above 86 km.
-    Furthermore, the *U.S. Standard Atmosphere 1976* model includes other gas
-    species than N2, O2, CO2 and CH4.
-    They are: Ar, He, Ne, Kr, H, O, Xe, He and H2.
-    All these species except H2 are absent from the
-    `HITRAN <https://hitran.org/>`_ spectroscopic database.
-    Since the absorption datasets are computed using HITRAN, the atomic species
-    could not be included in ``spectra-us76_u86_4``.
-    H2 was mistakenly forgotten and should be added to the dataset in a future
-    revision.
 
-
-    .. note::
-       We refer to the *U.S. Standard Atmosphere 1976* atmosphere model as the
-       model defined by the set of assumptions and equations in part 1 of the
-       report, and "numerically" illustrated by the extensive tables in part
-       4 of the report.
-       In particular, the part 3, entitled *Trace constituents*, which
-       provides rough estimates and discussions on the amounts of trace
-       constituents such as ozone, water vapor, nitrous oxide, methane, and so
-       on, is not considered as part of the *U.S. Standard Atmosphere 1976*
-       atmosphere model because it does not clearly defines the concentration
-       values of all trace constituents at all altitudes, neither does it
-       provide a way to compute them.
-
-    .. note::
-       It seems that the identifier "US76" is commonly used to refer to a
-       standard atmospheric profile used in radiative transfer applications.
-       However, there appears to be some confusion around the definition of
-       that standard atmospheric profile.
-       In our understanding, what is called the "US76 standard atmospheric
-       profile", or "US76" in short, **is not the U.S. Standard Atmosphere
-       1976 atmosphere model** but instead the so-called "U.S. Standard (1976)
-       atmospheric constituent profile model" in a AFGL technical report
-       entitled *AFGL Atmospheric Constituent Profiles (0-120km)* and
-       published in 1986 by Anderson et al
-       :cite:`Anderson1986AtmosphericConstituentProfiles`.
-       Although the "U.S. Standard (1976) atmospheric profile model" of the
-       AFGL's report is based on the *U.S. Standard Atmosphere* 1976 atmosphere
-       model (hence the name), it is significantly different when it comes
-       about the gas species concentration profiles.
-       Notably, the "U.S. Standard (1976) atmospheric profile model" of the
-       AFGL's report include radiatively active gases such as H2O, O3, N2O,
-       and CO, that the *U.S. Standard Atmosphere 1976* atmosphere model does
-       not include.
+    * Furthermore, the *U.S. Standard Atmosphere 1976* model includes other gas
+      species than N2, O2, CO2 and CH4.
+      They are: Ar, He, Ne, Kr, H, O, Xe, He and H2.
+      All these species except H2 are absent from the
+      `HITRAN <https://hitran.org/>`_ spectroscopic database.
+      Since the absorption datasets are computed using HITRAN, the atomic species
+      could not be included in ``spectra-us76_u86_4``.
+      H2 was mistakenly forgotten and should be added to the dataset in a future
+      revision.
+    
+    * We refer to the *U.S. Standard Atmosphere 1976* atmosphere model as the
+      model defined by the set of assumptions and equations in part 1 of the
+      report, and "numerically" illustrated by the extensive tables in part
+      4 of the report.
+      In particular, the part 3, entitled *Trace constituents*, which
+      provides rough estimates and discussions on the amounts of trace
+      constituents such as ozone, water vapor, nitrous oxide, methane, and so
+      on, is not considered as part of the *U.S. Standard Atmosphere 1976*
+      atmosphere model because it does not clearly defines the concentration
+      values of all trace constituents at all altitudes, neither does it
+      provide a way to compute them.
+    
+    * It seems that the identifier "US76" is commonly used to refer to a
+      standard atmospheric profile used in radiative transfer applications.
+      However, there appears to be some confusion around the definition of
+      that standard atmospheric profile.
+      In our understanding, what is called the "US76 standard atmospheric
+      profile", or "US76" in short, **is not the U.S. Standard Atmosphere
+      1976 atmosphere model** but instead the so-called "U.S. Standard (1976)
+      atmospheric constituent profile model" in a AFGL technical report
+      entitled *AFGL Atmospheric Constituent Profiles (0-120km)* and
+      published in 1986 by Anderson et al
+      :cite:`Anderson1986AtmosphericConstituentProfiles`.
+      Although the "U.S. Standard (1976) atmospheric profile model" of the
+      AFGL's report is based on the *U.S. Standard Atmosphere* 1976 atmosphere
+      model (hence the name), it is significantly different when it comes
+      about the gas species concentration profiles.
+      Notably, the "U.S. Standard (1976) atmospheric profile model" of the
+      AFGL's report include radiatively active gases such as H2O, O3, N2O,
+      and CO, that the *U.S. Standard Atmosphere 1976* atmosphere model does
+      not include.
     """
 
     _thermoprops: xr.Dataset = documented(
@@ -202,10 +202,7 @@ class US76ApproxRadProfile(RadProfile):
             return ureg.Quantity(np.zeros(profile.z_layer.size), "km^-1")
 
     def eval_sigma_a_ckd(self, *bindexes: Bindex, bin_set_id: str) -> pint.Quantity:
-        raise NotImplementedError(
-            "CKD data sets are not yet available for the U.S. Standard "
-            "Atmosphere 1976 atmopshere model."
-        )
+        raise UnsupportedModeError(supported="monochromatic")
 
     def eval_sigma_s_mono(self, w: pint.Quantity) -> pint.Quantity:
         profile = self.thermoprops
@@ -218,10 +215,7 @@ class US76ApproxRadProfile(RadProfile):
             return ureg.Quantity(np.zeros(profile.z_layer.size), "km^-1")
 
     def eval_sigma_s_ckd(self, *bindexes: Bindex) -> pint.Quantity:
-        wavelengths = ureg.Quantity(
-            np.array([bindex.bin.wcenter.m_as("nm") for bindex in bindexes]), "nm"
-        )
-        return self.eval_sigma_s_mono(w=wavelengths)
+        raise UnsupportedModeError(supported="monochromatic")
 
     def eval_albedo_mono(self, w: pint.Quantity) -> pint.Quantity:
         sigma_s = self.eval_sigma_s_mono(w)
@@ -231,19 +225,13 @@ class US76ApproxRadProfile(RadProfile):
         ).to("dimensionless")
 
     def eval_albedo_ckd(self, *bindexes: Bindex, bin_set_id: str) -> pint.Quantity:
-        sigma_s = self.eval_sigma_s_ckd(*bindexes)
-        sigma_t = self.eval_sigma_t_ckd(*bindexes, bin_set_id=bin_set_id)
-        return np.divide(
-            sigma_s, sigma_t, where=sigma_t != 0.0, out=np.zeros_like(sigma_s)
-        ).to("dimensionless")
+        raise UnsupportedModeError(supported="monochromatic")
 
     def eval_sigma_t_mono(self, w: pint.Quantity) -> pint.Quantity:
         return self.eval_sigma_a_mono(w) + self.eval_sigma_s_mono(w)
 
     def eval_sigma_t_ckd(self, *bindexes: Bindex, bin_set_id: str) -> pint.Quantity:
-        return self.eval_sigma_a_ckd(
-            *bindexes, bin_set_id=bin_set_id
-        ) + self.eval_sigma_s_ckd(bindexes)
+        raise UnsupportedModeError(supported="monochromatic")
 
     def eval_dataset_mono(self, w: pint.Quantity) -> xr.Dataset:
         profile = self.thermoprops
@@ -256,13 +244,4 @@ class US76ApproxRadProfile(RadProfile):
         ).squeeze()
 
     def eval_dataset_ckd(self, *bindexes: Bindex, bin_set_id: str) -> xr.Dataset:
-        if len(bindexes) > 1:
-            raise NotImplementedError
-        else:
-            return make_dataset(
-                wavelength=bindexes[0].bin.wcenter,
-                z_level=to_quantity(self.thermoprops.z_level),
-                z_layer=to_quantity(self.thermoprops.z_layer),
-                sigma_a=self.eval_sigma_a_ckd(*bindexes, bin_set_id=bin_set_id),
-                sigma_s=self.eval_sigma_s_ckd(*bindexes),
-            ).squeeze()
+        raise UnsupportedModeError(supported="monochromatic")
