+++ tests/02_eradiate/01_unit/scenes/atmosphere/test_molecular_atmosphere.py
@@ -3,6 +3,7 @@
 import numpy as np
 import pytest
 
+import eradiate
 from eradiate import path_resolver
 from eradiate.contexts import KernelDictContext
 from eradiate.scenes.atmosphere import MolecularAtmosphere
@@ -52,18 +53,17 @@ def afgl_1986_test_absorption_data_sets():
     }
 
 
-def test_molecular_atmosphere_afgl_1986(
-    mode_mono,
-    tmpdir,
-    afgl_1986_test_absorption_data_sets,
-):
+@pytest.mark.parametrize(
+    "bin", ["280", "550", "1040", "2120", "2400"]
+)
+def test_molecular_atmosphere_afgl_1986(mode_ckd, bin):
     """MolecularAtmosphere 'afgl_1986' constructor produces a valid kernel
     dictionary."""
-    ctx = KernelDictContext(spectral_ctx={"wavelength": 550.0})
-    atmosphere = MolecularAtmosphere.afgl_1986(
-        geometry="plane_parallel",
-        absorption_data_sets=afgl_1986_test_absorption_data_sets,
-    )
+    bin = eradiate.scenes.measure._core.CKDMeasureSpectralConfig(bins=bin).bins[0]
+    bindex = eradiate.ckd.Bindex(bin=bin, index=3)
+    spectral_ctx = eradiate.contexts.CKDSpectralContext(bindex=bindex, bin_set="10nm")
+    ctx = KernelDictContext(spectral_ctx=spectral_ctx)
+    atmosphere = MolecularAtmosphere.afgl_1986(geometry="plane_parallel")
     assert atmosphere.kernel_dict(ctx).load()
